Refactoring of core package structure
